Updated dagger setup
