+++ app/build.gradle
@@ -4,7 +4,7 @@ apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
 apply plugin: 'androidx.navigation.safeargs'
 
-def dagger_version = 2.20
+def dagger_version = 2.23
 def nav_version = "2.1.0-alpha04"
 
 android {
