+++ app/src/main/java/com/jazart/symphony/repository/posts/FirebaseOfflinePostDataSource.kt
@@ -6,14 +6,15 @@ import com.jazart.symphony.common.Constants
 import com.jazart.symphony.common.Constants.AUTHOR
 import com.jazart.symphony.repository.AbstractFirebaseDataSource
 import com.jazart.symphony.repository.await
+import dagger.Reusable
 import entities.Post
 import javax.inject.Inject
 import javax.inject.Singleton
 
-@Singleton
-class FirebaseOfflinePostDataSource @Inject constructor(source: Source) : AbstractFirebaseDataSource(source), PostRepository {
-
-    constructor(): this(Source.CACHE)
+@Reusable
+class FirebaseOfflinePostDataSource constructor(source: Source) : AbstractFirebaseDataSource(source), PostRepository {
+    @Inject
+    constructor() : this(Source.CACHE)
 
     override suspend fun loadPostById(id: String): Post? {
         return db.collection(Constants.POSTS).document(id).get(source).await().toObject(Post::class.java)
