+++ app/src/main/java/com/jazart/symphony/repository/songs/FirebaseOnlineSongDataSource.kt
@@ -6,13 +6,16 @@ import com.jazart.symphony.common.Constants.AUTHOR
 import com.jazart.symphony.common.Constants.SONGS
 import com.jazart.symphony.repository.AbstractFirebaseDataSource
 import com.jazart.symphony.repository.await
+import dagger.Reusable
 import entities.Song
 import javax.inject.Inject
 import javax.inject.Singleton
 
-@Singleton
-class FirebaseOnlineSongDataSource @Inject constructor(source: Source) : AbstractFirebaseDataSource(source), SongRepository {
-    constructor(): this(Source.SERVER)
+@Reusable
+class FirebaseOnlineSongDataSource constructor(source: Source) : AbstractFirebaseDataSource(source), SongRepository {
+    @Inject
+    constructor() : this(Source.SERVER)
+
     override suspend fun findSongById(id: String): Song? {
         return db.collection(SONGS).document(id).get(source).await().toObject(Song::class.java)
     }
