+++ app/src/main/java/com/jazart/symphony/repository/songs/SongRepositoryImpl.kt
@@ -5,16 +5,17 @@ import com.jazart.data.repo.SongRepository
 import com.jazart.symphony.repository.InMemoryDataSource
 import com.jazart.symphony.repository.posts.AbstractRepository
 import com.jazart.symphony.repository.users.FetchStrategy
+import dagger.Reusable
 import entities.Song
 import javax.inject.Inject
 import javax.inject.Singleton
 
-@Singleton
+@Reusable
 class SongRepositoryImpl @Inject constructor(connection: ConnectivityManager,
-                         memory: InMemoryDataSource<Song>,
-                         private val disk: FirebaseOfflineSongDataSource,
-                         private val network: FirebaseOnlineSongDataSource,
-                         fetchStrategy: FetchStrategy = FetchStrategy.NETWORK_FIRST) : AbstractRepository<Song>(connection, memory), SongRepository {
+                                             @Singleton memory: InMemoryDataSource<Song>,
+                                             private val disk: FirebaseOfflineSongDataSource,
+                                             private val network: FirebaseOnlineSongDataSource,
+                                             fetchStrategy: FetchStrategy = FetchStrategy.NETWORK_FIRST) : AbstractRepository<Song>(connection, memory), SongRepository {
 
     override suspend fun findSongById(id: String): Song? {
         return super.load(id, { disk.findSongById(id) }) { network.findSongById(id) }
