+++ app/src/main/java/com/jazart/symphony/repository/users/UserRepositoryImpl.kt
@@ -1,17 +1,21 @@
 package com.jazart.symphony.repository.users
 
 import android.net.ConnectivityManager
-import com.jazart.symphony.repository.posts.AbstractRepository
+import com.jazart.data.repo.UserRepository
 import com.jazart.symphony.repository.InMemoryDataSource
+import com.jazart.symphony.repository.posts.AbstractRepository
+import dagger.Reusable
 import entities.User
 import javax.inject.Inject
+import javax.inject.Singleton
 
+@Reusable
 class UserRepositoryImpl @Inject constructor(connection: ConnectivityManager,
-                                             memory: InMemoryDataSource<User>,
+                                             @Singleton memory: InMemoryDataSource<User>,
                                              private val disk: FirebaseOfflineUserDataSource,
                                              private val network: FirebaseOnlineUserDataSource,
                                              fetchStrategy: FetchStrategy = FetchStrategy.NETWORK_FIRST) :
-        AbstractRepository<User>(connection, memory), com.jazart.data.repo.UserRepository {
+        AbstractRepository<User>(connection, memory), UserRepository {
 
     override suspend fun getUserById(id: String): User? {
         return super.load(id, { disk.getUserById(id) }, { network.getUserById(id) })
